Get image IO working
Looks like I have fiiinally managed to get the image file IO side of things working, while possibly also doing histogram equalization on the output image.

+++ rust/src/common.rs
@@ -2,9 +2,9 @@
 pub struct Parameters {
     pub left_image: Vec<u8>,
     pub right_image: Vec<u8>,
-    pub width: usize,
-    pub height: usize,
-    pub total_pixels: usize,
+    pub width: u32,
+    pub height: u32,
+    pub total_pixels: u32,
     pub window_edge_size: usize,
     pub maximum_disparity: usize,
     pub use_zero_mean: bool,
